refactor(types): вынести ServiceModuleEventRegistrar из ConfigurationTypesProvider
CTP перевалил за 750 строк (Sonar S104). Развёртка generic-events
HTTP/Web/Integration сервисных модулей логически отделима и вынесена в
отдельный @WorkspaceScope bean — CTP только инжектит и вызывает.

Прочие фиксы:
- EventHandlerInvalidSignatureDiagnostic.matchingRparenIndex: if-else if на
  switch expression (Sonar S6880)
- ConfigurationTypesRegisteredEvent: @Scope(singleton) — подавить S6832
  без @SuppressWarnings (запрещено правилом S1309)
